refactor(router): move route config stuff into its own folder
