+++ modules/angular2/src/router/route_config/route_config_decorator.ts
