+++ modules/angular2/src/router/route_config/route_config_nomalizer.ts
@@ -1,8 +1,8 @@
 import {AsyncRoute, AuxRoute, Route, Redirect, RouteDefinition} from './route_config_decorator';
-import {ComponentDefinition} from './route_definition';
+import {ComponentDefinition} from '../route_definition';
 import {isType, Type} from 'angular2/src/facade/lang';
 import {BaseException, WrappedException} from 'angular2/src/facade/exceptions';
-import {RouteRegistry} from './route_registry';
+import {RouteRegistry} from '../route_registry';
 
 
 /**
