+++ modules/angular2/src/router/rules/rule_set.ts
@@ -10,7 +10,7 @@ import {
   RouteMatch,
   PathMatch
 } from './rules';
-import {Route, AsyncRoute, AuxRoute, Redirect, RouteDefinition} from '../route_config_impl';
+import {Route, AsyncRoute, AuxRoute, Redirect, RouteDefinition} from '../route_config/route_config_impl';
 import {AsyncRouteHandler} from '../route_handlers/async_route_handler';
 import {SyncRouteHandler} from '../route_handlers/sync_route_handler';
 import {Url} from '../url_parser';
