+++ modules/angular2/test/router/integration/lifecycle_hook_spec.ts
@@ -31,7 +31,7 @@ import {
   AuxRoute,
   AsyncRoute,
   Redirect
-} from 'angular2/src/router/route_config_decorator';
+} from 'angular2/src/router/route_config/route_config_decorator';
 
 import {
   OnActivate,
