+++ modules/angular2/test/router/rule_set_spec.ts
@@ -15,7 +15,7 @@ import {Map, StringMapWrapper} from 'angular2/src/facade/collection';
 import {RouteMatch, PathMatch, RedirectMatch} from 'angular2/src/router/rules/rules';
 import {RuleSet} from 'angular2/src/router/rules/rule_set';
 
-import {Route, Redirect} from 'angular2/src/router/route_config_decorator';
+import {Route, Redirect} from 'angular2/src/router/route_config/route_config_decorator';
 import {parser} from 'angular2/src/router/url_parser';
 import {PromiseWrapper} from 'angular2/src/facade/promise';
